requirements.txt
sqlalchemy==1.3.19
pandas==0.25.3
pyyaml==5.3.1
mysqlclient==2.0.1
mysql-connector-python==8.0.21
xlrd==1.2.0
